refactor(speech): split voice_loop state_machine.rs into responsibility submodules
The 786-line state_machine.rs becomes a directory-style module split by
responsibility (the state_machine/ dir already existed for mocks.rs):

- state_machine/mod.rs (74) — module doc, the flat `pub use` facade
  (9-symbol external surface unchanged), and the private FALLBACK_LINE +
  handle_llm_err. These stay in mod.rs because mocks.rs (the test module,
  a child of this module) reaches FALLBACK_LINE via `super::`, and a
  #[cfg(test)]-gated observer-type import keeps mocks.rs's
  `super::ExitReason`/`TurnCompletePayload`/`VoiceState` paths resolving
  with zero test-file churn.
- state_machine/types.rs (189) — LoopConfig, LoopBackends (+ impl),
  Responder, LoopHandle, DrainHook, VoiceLoopError,
  VAD_EVENT_CHANNEL_CAPACITY (verbatim).
- state_machine/entry.rs (90) — run_loop + run_loop_borrowed (verbatim).
- state_machine/inner.rs (470) — run_loop_inner, now pub(super)
  (verbatim body; imports observer/quit_detect/tts_markdown via
  crate::voice_loop:: paths instead of the old parent-relative super::).
- state_machine/mocks.rs — untouched.

Behaviour-preserving proof:
- cargo test -p primer-speech --features voice-loop: 101 passed /
  0 failed / 2 ignored (lib) + 1 + 10 integration — identical to the
  pre-change baseline.
- External pub-surface diff old vs new: empty.
- cargo clippy -p primer-speech --features voice-loop,silero,whisper,piper
  --all-targets -- -D warnings: clean (cfg-arm dual-verify).
- cargo clippy -p primer-speech --all-targets -- -D warnings: clean.
- cargo fmt --check: clean. Full-workspace cargo test: green.

CLAUDE.md gets a module-layout note on the primer-speech bullet
(mirrors the #304/#305 precedent).
